+++ lib/comic_book/cbr.rb
@@ -1,4 +1,5 @@
 require 'shellwords'
+require 'comicinfo'
 require_relative 'adapter'
 require_relative 'cbr/extractor'
 
@@ -10,6 +11,19 @@ class ComicBook
 
     def extract options = {}
       Extractor.new(path).extract options
+    end
+
+    def info
+      entries = CLIHelpers.lsar_list path
+      return nil unless entries.include? 'ComicInfo.xml'
+
+      Dir.mktmpdir do |temp_dir|
+        CLIHelpers.unar_extract path, temp_dir
+        xml_path = File.join temp_dir, 'ComicInfo.xml'
+        return nil unless File.exist? xml_path
+
+        ComicInfo.load xml_path
+      end
     end
 
     def pages